add bfloat16 support and set TORCH_CUDNN_V8_API_ENABLED

+++ bsrt/metrics/aligned_l2.py
@@ -56,7 +56,7 @@ class AlignedL2(Metric):
 
         # TODO: The generated superresolution image regularly has a range greater than 1.0. Is this a problem?
         ssim: Tensor = compute_ssim(
-            pred_warped_m.contiguous(),
+            pred_warped_m.type_as(gt).contiguous(),
             gt.contiguous(),
             gaussian_kernel=True,
             kernel_size=11,
